pyproject.toml
[tool.poetry]
name = "tfresurrect"
version = "0.1.0"
description = "resurrect terraform state"
authors = ["Kapil Thangavelu <kapilt@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.8"
python-hcl2 = "^2.0.3"
pytest-terraform = "^0.5.3"
c7n = "^0.9.11"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.6b0"
isort = "^5.8.0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
